Start of a Scala port.
So far, this has support for:

- Parsing Dedalus programs
- Rewriting programs to reference clock relations
- Generating provenance-tracking rules
- Type inference
- C4 input and output

I still need to implement derivation tree construction and
SAT input generation.  There's also work to be done in the
clock fact generator.  I also need to add generation of
Dedalus-style persistence rules.

I'd also like to change how we handle include paths in Dedalus
programs: right now, I assume the include path is relative to
the absolute path of the first input file, so you might need to
change the include paths in the examples in order to get them
to run.

I also need to add tests.