ProvenanceReader: ensure negated tuples weren't derived.
This commit adds some sanity checks to the provenance rules.  If a
provenance table says that a rule containing a notin fired, then there
shouldn't be any derivations that negate the notin.

This is implemented as two separate checks.  For a negated predicate
like `notin pred(.)`, we search the `pred` table for any matching
tuples.  We also check that there are no matching entries in `pred`'s
provenance-tracking tables.  If we find a match in either set of tables,
then something has gone wrong with the provenance tracking rules.

While searching for the negated tuple, we also check that a `pred` tuple
has a matching provenance table entry _if and only if_ it appears in the
`pred` table; if this is false, then something has gone horribly wrong.