Update tests due to different semantics of EFF in Scala version:
In the Scala verifier, all omissions and crash failures must occur
_before_ the EFF.  In the Ruby verifier, crash failures are allowed to
to occur _at_ the EFF.  This was causing spurious failures in the new
counterexample suite.

For example, the canonical counterexample for rdlog with crash failures
is to drop a single message at t = 1, then crash the sender at t = 2. 
If EFF = 2, then the verifier won't consider this failure scenario and
will fail to find a counterexample.

To address this, I just bumped up EFF from 2 to 3 in the tests.  All of
those tests are passing now.

Also, fix an off-by-one error in a debug message that's printed while
installing clock rules.