Fix provenance rules for aggregates:
Given a rule like

	some_count(X, count<Y>) :- a(X, Z), b(Z, Y)

the old rewrite recorded 'Z' in the head because it was bound in the
rule body; this is wrong because it affects the grouping and changes the
aggregate result.

Now, prior to applying the provenance rewrites, we split each aggregate
rule into two rules, one that records the variable bindings and another
that performs the actual aggregation:

	some_count_vars(X, Y, Z) :- a(X, Z), b(Z, Y)
	some_count(X, count<Y>) :- some_count_vars(X, Y, _)

Now, the provenance reconstruction is trivial: when searching for the
provenance of a tuple in `some_count`, we can easily determine that
the rule firing depends on every matching tuple in `some_count_vars`.