Remove redundant wildcard goals from agg rule nodes.
We were adding additional subgoals for all of the tuples that
contributed to the aggregate, but also including the original goal
that contained wildcards, leading to a massive explosion of equivalent proofs.

For example, when scanning for the provenance of `foo("a", 3)` using the rule

	foo(X, count<I>) :- bar(X, _, I), baz(X)

we might construct a rule firing with subgoals

	baz("a")
	bar("a", _, _),
	bar("a", "z", 1),
	bar("a", "x", 4),
	bar("a", "y", 2)

Although each concrete `bar` tuple might only have a single derivation, the
wildcard `bar("a", _, _)` has three derivations.  This caused problems when
mining proof trees and caused Paxos to exhaust the verifier's memory.

The solution in this patch ensures that we never create wildcard subgoals if
those subgoals contribute to aggregates.