Write harness results into single CSV file.
If we care about specific metrics, we can pull them out of the registry
and add them as extra CSV columns.  Any other metrics are treated as
semistructured data and dumped into a JSON-valued "metrics" column.
This ensures that we record all information captured in each run, while
allowing us to provide easier spreadsheet access to the most important
metrics.