Fix type inference for variables appearing in quals.
Before, we assumed that any variable appearing in a qual had to be an
integer.  Instead, we now treat quals as evidence that the two columns
must have the same type.

The current implementation assumes that every variable appearing in a
qual must have the same type.  In a rich enough expression language,
this wouldn’t necessarily be true, since you could have expressions
that contain logical connectives.  I expect that most Dedalus programs
will use multiple quals for AND and multiple rules for OR, so I think
we can get away with this simplistic type-checking / inference.