Fix Z3 solver bugs; add support for crashes.
The logic for goals/rules was mixed up: a goal node’s boolean expression should be true if ALL rule firings are invalidated, and a rule firing
is invalidated if ANY of its subgoals is invalidated.

The crash failure support uses pseudo-boolean constraints to express
the “at-least-k nodes don’t crash” requirement.